% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cor.R
\name{gen_cor}
\alias{gen_cor}
\title{Generate Heatmaps}
\usage{
gen_cor(obj, gl, tissues, genes, exp, samples, filePrefix = "file",
  thresh = 0.9, dosave = T, meancor = 0.1, ...)
}
\arguments{
\item{obj}{an teQTL class object that has been initialized using "init_teQTL"}

\item{gl}{named character vector of ensembl gene IDs. Names should be human readable gene-names}

\item{tissues}{character vector of tissues (from the samples data.frame in doi: 10.5281/zendo.838733) specifying those tissues of where the regression should be performed (not yet implimented). If no tissues are specified, the default is all tissues.}

\item{genes}{gene annotation data.frame from doi: 10.5281/zendo.838733}

\item{exp}{gene expression matrix from doi: 10.5281/zendo.838733}

\item{samples}{samples data.frame in doi: 10.5281/zendo.838733}

\item{filePrefix}{string, filename for the output}

\item{thresh}{number [0,1], minimum cummulative probability of TF occcurances. default is .9}

\item{dosave}{boolean, if true, the regression visualization will be saved in the location specified by the filePrefix}

\item{meancor}{number, minimum correlation to be included in the visualization}

\item{...}{parameters to be passed to the pdf function e.g. height and width}
}
\value{
object of class "init_teQTL" extended to contain gene expression correlation between Targets and TFs in obj$cr_TSspec (only specified tissues) and obj$cr_all (all tissues) respectively
}
\description{
Generate heatmaps  of trans-eQTL likelihood: -log10( cummumaltive probability of number of hits for general or specific TF)
}
