% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_heatmap.R
\name{gen_heatmap}
\alias{gen_heatmap}
\title{Generate Heatmaps}
\usage{
gen_heatmap(obj, gl, filePrefix = "file", thresh = 0.9, dosave = T,
  minNeglog10Perc = 1, ...)
}
\arguments{
\item{obj}{an teQTL class object that has been initialized using "init_teQTL"}

\item{gl}{named character vector of ensembl gene IDs. Names should be human readable gene-names}

\item{filePrefix}{string, filename for the output}

\item{thresh}{number [0,1], minimum cummulative probability of TF occcurances. default is .9}

\item{dosave}{boolean, if true, the regression visualization will be saved in the location specified by the filePrefix}

\item{...}{parameters to be passed to the pdf function e.g. height and width}

\item{min}{number, minimum -log10(cummulative probability of TF occcurances), default is 1}
}
\value{
object of class "init_teQTL" extended to contain TF specific and general aggregated -log10 of cummulative probabilities for TF hit counts for each target stored in obj$TF_v_targets.cdf_all and obj$TF_v_targets.cdf_spec respectively
}
\description{
Generate heatmaps  of trans-eQTL likelihood: -log10( cummumaltive probability of number of hits for general or specific TF)
}
