% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_regression.R
\name{gen_regression}
\alias{gen_regression}
\title{Generate Regressions}
\usage{
gen_regression(obj, gl, tissues = NULL, genes, exp, samples,
  filePrefix = "file", dosave = T, min = 2)
}
\arguments{
\item{obj}{an teQTL class object that has been initialized using "init_teQTL"}

\item{gl}{named character vector of ensembl gene IDs. Names should be human readable gene-names}

\item{tissues}{character vector of tissues (from the samples data.frame in doi: 10.5281/zendo.838733) specifying those tissues of where the regression should be performed (not yet implimented). If no tissues are specified, the default is all tissues.}

\item{genes}{gene annotation data.frame from doi: 10.5281/zendo.838733}

\item{exp}{gene expression matrix from doi: 10.5281/zendo.838733}

\item{samples}{samples data.frame in doi: 10.5281/zendo.838733}

\item{filePrefix}{string, filename for the output}

\item{dosave}{boolean, if true, the regression visualization will be saved in the location specified by the filePrefix}

\item{min}{number, minimum -log10(cummulative probability of TF occcurances), default is 2}
}
\value{
object of class "init_teQTL" extended to contain regression objects for each target stored in obj$regressions and visualizations of those regressions stored in obj$regression_fig
}
\description{
Takes predicted trans-eQTLs and fits regressions to describe the relation between predicted TFs and the target
}
