% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_teQTL.R
\name{init_teQTL}
\alias{init_teQTL}
\title{Initialize the teQTL object}
\usage{
init_teQTL(gl, tissues, edges, genes, netTS, expTS, minTS = 0, rand = 50,
  onlyCanonical = T, filePrefix = "file", thresh = 0.9, dosave = T)
}
\arguments{
\item{gl}{named character vector of ensembl gene IDs. Names should be human readable gene-names}

\item{tissues}{character vector of tissues (from the samples data.frame in doi: 10.5281/zendo.838733) specifying those tissues of where the regression should be performed (not yet implimented). If no tissues are specified, the default is all tissues.}

\item{edges}{edges data.frame containing predicted trans-eQTL interactions from doi: 10.5281/zendo.838733}

\item{genes}{gene annotation data.frame from doi: 10.5281/zendo.838733}

\item{netTS}{netTS data.frame in doi: 10.5281/zendo.838733}

\item{expTS}{expTS data.frame in doi: 10.5281/zendo.838733}

\item{minTS}{integer, minimum number of specified tissues which must contain a teQTL for it to be considered. default is 0 which disables this parameter because it is too limiting.}

\item{rand}{inteber, number of random genelists to examine for the background}

\item{filePrefix}{string, filename for the output}

\item{thresh}{number [0,1], minimum cummulative probability of TF occcurances. default is .9}

\item{dosave}{boolean, if true, the regression visualization will be saved in the location specified by the filePrefix}
}
\value{
object of class "init_teQTL" extended to contain TF specific and general cummulative probabilities for TF hit counts for each target stored in obj$TF_sel
}
\description{
Initilizes the fundamental teQTL object
}
