% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_fig.R
\name{regression_fig}
\alias{regression_fig}
\title{Generate Regression Figure}
\usage{
regression_fig(obj, genes)
}
\arguments{
\item{obj}{regression object output by $gen_regression$ function}

\item{genes}{gene annotation data.frame from doi: 10.5281/zendo.838733}
}
\value{
grid.arrange output of a list of ggplot grobs
}
\description{
Takes regression objects and visualizes them
}
